grid: Rewrite allocation algorithm
The 'pushing aside' effect is no longer desired, so take this opportunity
to rewrite the allocation algorithm and massively simplify it. Instead of
storing data on each row, just store a single variable that represents
where the current fully-expanded row is. This is a floating-point number,
and is used to determine the scale of the height for each row.

As expanded children now overlap other children, this introduced a problem
with overlapping shadows when we're trying to clip around the focused
child. To work around this, I've made it so that the shadow drawn under
a child is changed when that child is focused. We may want to revise this
somehow, perhaps with some special-casing...

The tile-width and tile-height properties are now convenience properties
for the size that a tile is expected to occupy, and are only notified on
outside of allocation - this allows users to bind these properties to
size related properties of the children inside the grid.